Add files via upload
This is a website where I used some bootstrap and Javascript to implement some functions. Open the HomeVersieB file, you'll see a screen with a button. If you click on the button some pop-ups pop up and you can proceed to the next website where you can give your name and some code. After you have put in the information and clicked 'OK' it will tell you if the code is right or not. If you put in 123 it will tell you that the code is right if not then it will tell you its the wrong code.